requirements.txt
python-telegram-bot==21.3
playwright==1.44.0
python-dotenv==1.0.1
aiohttp==3.9.5
aiofiles==23.2.1
